py-django-crewai/utils/favicon_requirements.txt
cairosvg==2.7.1
Pillow==11.2.1
